fix: resolve Math.Round ambiguity by casting Seat.Price to decimal before rounding

+++ Models/Booking/SeatDto.cs
@@ -4,6 +4,6 @@
     {
         public string MaGhe { get; set; }
 
-        public decimal Price { get; set; }
+        public int Price { get; set; } // ✅ ĐÃ SỬA
     }
 }
